Add support for locking data in memory with mlock
This commit adds the ability to lock data in memory to prevent it from being swapped to disk using the mlock syscall. The mlock function is called on the memory-mapped data after it is opened. If mlock fails, the program logs an error message with instructions on how to fix the issue. Additionally, some of the functions that accept string arguments now accept byte slice arguments instead for better performance.